started stage 2 of the prototype
halfway through the prototype and realized i haven't been commiting (classic)
a small recap:
tweaked some aspects on the first level, still need to change the octopus to a wall.

the overlap is VERY annoying in the fact that it overlaps to far away from the user to the wall

started the second stage by adding the colored walls, one of the 4 walls has been overlaped.